+++ tests/layer_validate.rs
@@ -101,6 +101,18 @@ fn validate_doctl() {
 
 #[test]
 #[ignore]
+fn validate_ddog() {
+    validate("ddog");
+}
+
+#[test]
+#[ignore]
+fn validate_sntry() {
+    validate("sntry");
+}
+
+#[test]
+#[ignore]
 fn validate_postgres() {
     validate("postgres");
 }
@@ -111,6 +123,12 @@ fn validate_postgres() {
 #[ignore]
 fn validate_lin() {
     validate("lin");
+}
+
+#[test]
+#[ignore]
+fn validate_secunit() {
+    validate("secunit");
 }
 
 #[test]